# Refactor to Remove Config Manager and Optimize Settings