--- a/.gitignore
+++ b/.gitignore
@@ -10,7 +10,6 @@ __pycache__/
 *$py.class
 chroma_db/
 .vscode
-.project_doc_record
 # C extensions
 *.so
 
@@ -177,7 +176,6 @@ prompt_output/
 .vscode/
 
 # RepoAgent
-.project_doc_record
 log.txt
 .chroma_db
 config.yml