--- a/display/book_tools/generate_repoagent_books.py
+++ b/display/book_tools/generate_repoagent_books.py
@@ -1,6 +1,5 @@
 import os
 import shutil
-
 import sys