--- a/repo_agent/chat_engine.py
+++ b/repo_agent/chat_engine.py
@@ -1,4 +1,4 @@
-from llama_index.llms.openai_like import OpenAILike
+from llama_index.llms.openai import OpenAI
 
 from repo_agent.doc_meta_info import DocItem
 from repo_agent.log import logger
@@ -14,7 +14,7 @@ class ChatEngine:
     def __init__(self, project_manager):
         setting = SettingsManager.get_setting()
 
-        self.llm = OpenAILike(
+        self.llm = OpenAI(
             api_key=setting.chat_completion.openai_api_key.get_secret_value(),
             api_base=setting.chat_completion.openai_base_url,
             timeout=setting.chat_completion.request_timeout,
@@ -119,12 +119,12 @@ class ChatEngine:
 
         try:
             response = self.llm.chat(messages)
-            logger.debug(f"LLM Prompt Tokens: {response.raw.usage.prompt_tokens}")  # type: ignore #
+            logger.debug(f"LLM Prompt Tokens: {response.raw.usage.prompt_tokens}")  # type: ignore
             logger.debug(
-                f"LLM Completion Tokens: {response.raw.usage.completion_tokens}"  # type: ignore #
+                f"LLM Completion Tokens: {response.raw.usage.completion_tokens}"  # type: ignore
             )
             logger.debug(
-                f"Total LLM Token Count: {response.raw.usage.total_tokens}"  # type: ignore #
+                f"Total LLM Token Count: {response.raw.usage.total_tokens}"  # type: ignore
             )
             return response.message.content
         except Exception as e: