--- a/repo_agent/chat_with_repo/json_handler.py
+++ b/repo_agent/chat_with_repo/json_handler.py
@@ -32,7 +32,7 @@ class JsonFileProcessor:
                     if "md_content" in item and item["md_content"]:
                         # Append the first element of 'md_content' to the result list
                         md_contents.append(item["md_content"][0])
-                          # Build a dictionary containing the required information
+                        # Build a dictionary containing the required information
                         item_dict = {
                             "type": item.get("type", "UnknownType"),
                             "name": item.get("name", "Unnamed"),
@@ -45,7 +45,7 @@ class JsonFileProcessor:
                             # Adapt or remove fields based on new structure requirements
                         }
                         extracted_contents.append(item_dict)
-        return md_contents,extracted_contents
+        return md_contents, extracted_contents
 
     def recursive_search(self, data_item, search_text, code_results, md_results):
         if isinstance(data_item, dict):
@@ -53,22 +53,22 @@ class JsonFileProcessor:
             for key, value in data_item.items():
                 # Recursively search through dictionary values and lists
                 if isinstance(value, (dict, list)):
-                    self.recursive_search(value, search_text,code_results, md_results)
+                    self.recursive_search(value, search_text, code_results, md_results)
         elif isinstance(data_item, list):
             for item in data_item:
                 # Now we check for the 'name' key in each item of the list
-                if isinstance(item, dict) and item.get('name') == search_text:
+                if isinstance(item, dict) and item.get("name") == search_text:
                     # If 'code_content' exists, append it to results
-                    if 'code_content' in item:
-                        code_results.append(item['code_content'])
-                        md_results.append(item['md_content'])
+                    if "code_content" in item:
+                        code_results.append(item["code_content"])
+                        md_results.append(item["md_content"])
                 # Recursive call in case of nested lists or dicts
                 self.recursive_search(item, search_text, code_results, md_results)
 
     def search_code_contents_by_name(self, file_path, search_text):
         # Attempt to retrieve code from the JSON file
         try:
-            with open(file_path, 'r', encoding='utf-8') as file:
+            with open(file_path, "r", encoding="utf-8") as file:
                 data = json.load(file)
                 code_results = []
                 md_results = []  # List to store matching items' code_content and md_content
@@ -88,4 +88,4 @@ class JsonFileProcessor:
 
 if __name__ == "__main__":
     processor = JsonFileProcessor("database.json")
-    md_contents,extracted_contents = processor.extract_data()
+    md_contents, extracted_contents = processor.extract_data()